+++ example/definition-files-absolute-paths/src/main.ts
@@ -1,3 +1,4 @@
 import { Model } from 'models';
+import { MyClass } from 'MyClass';
 
 export class UnusedClassFromMain implements Model {}