+++ src/app.ts
@@ -6,29 +6,35 @@ import { dirname, resolve } from 'path';
 import { TsConfig } from './types';
 import { extractOptionsFromFiles } from './argsParser';
 import parseFiles from './parser';
-import { readFileSync } from 'fs';
 
 const parseTsConfig = (tsconfigPath: string): TsConfig => {
   const basePath = resolve(dirname(tsconfigPath));
 
   try {
-    const parseJsonResult = ts.parseConfigFileTextToJson(
+    const configFileName = ts.findConfigFile(
+      basePath,
+      ts.sys.fileExists,
       tsconfigPath,
-      readFileSync(tsconfigPath, { encoding: 'utf8' }),
     );
+    if (!configFileName) throw `Couldn't find ${tsconfigPath}`;
 
-    if (parseJsonResult.error) throw parseJsonResult.error;
+    const configFile = ts.readConfigFile(configFileName, ts.sys.readFile);
 
     const result = ts.parseJsonConfigFileContent(
-      parseJsonResult.config,
+      configFile.config,
       ts.sys,
       basePath,
+      undefined,
+      tsconfigPath,
     );
     if (result.errors.length) throw result.errors;
 
+    // We now use absolute paths to avoid ambiguity and to be able to delegate baseUrl resolving to TypeScript.
+    // A consequence is, we cannot fall back to '.' so instead the fallback is the tsconfig dir:
+    // (I think this only occurs with unit tests!)
     return {
-      baseUrl: result.raw?.compilerOptions?.baseUrl || '.',
-      paths: result.raw?.compilerOptions?.paths,
+      baseUrl: result.options.baseUrl || basePath,
+      paths: result.options.paths,
       files: result.fileNames,
     };
   } catch (e) {
@@ -53,8 +59,5 @@ export default (tsconfigPath: string, files?: string[]): Analysis => {
   const args = extractOptionsFromFiles(files);
   const tsConfig = loadTsConfig(tsconfigPath, args.tsFiles);
 
-  return analyze(
-    parseFiles(dirname(tsconfigPath), tsConfig, args.options),
-    args.options,
-  );
+  return analyze(parseFiles(tsConfig, args.options), args.options);
 };